bug fixes, and cleanup
