+++ pyhealth/interpret/__init__.py
@@ -1,3 +1,1 @@
-# from pyhealth.interpret import methods 
-# from pyhealth.interpret import methods
-from pyhealth.interpret import tasks
+from pyhealth.interpret import methods
