+++ pyhealth/interpret/methods/__init__.py
@@ -1,0 +1,1 @@
+from pyhealth.interpret.methods.chefer import CheferRelevance
