--- pyhealth/interpret/tasks/__init__.py
@@ -1,2 +1,0 @@
-# from pyhealth.interpret.tasks.medical_code import MedicalCodeInterpreter
-# from pyhealth.interpret.tasks.vision import VisionInterpreter
