--- pyhealth/interpret/tasks/eeg.py
@@ -1,5 +1,0 @@
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-
-# to be implemented with PyHealth
