--- pyhealth/interpret/tasks/methods/__init__.py
@@ -1,1 +1,0 @@
-from pyhealth.interpret.tasks.methods.chefer import CheferRelevance
